requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.5
django-celery-results==2.5.1
django-channels==0.7.0
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
gunicorn==21.2.0
h11==0.14.0
idna==3.4
kombu==5.3.2
oauthlib==3.2.2
packaging==23.1
prompt-toolkit==3.0.39
psycopg2==2.9.7
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.4
uvicorn==0.23.2
vine==5.0.0
wcwidth==0.2.6
